std: split spec definitions
